did suicide behaviour

+++ Gooseberries/Assets/Scripts/AI/ScriptableObject/EnemyStatsData.cs
@@ -7,7 +7,7 @@ public class EnemyStatsData : ScriptableObject
 {
     [field: SerializeField] public float shootCooldown { get; private set; }
     [field: SerializeField] public int ammo {get; private set;}
-    [field: SerializeField] public float arrowDmg {get; private set;}
-    [field: SerializeField] public float suicideDmg {get; private set;}
+    [field: SerializeField] public int arrowDmg {get; private set;}
+    [field: SerializeField] public int suicideDmg {get; private set;}
     [field: SerializeField] public float disableDuration {get; private set;}
 }
